respeaker/requirements.txt
#
####### requirements.txt #######
#
###### Respeaker Home Automation Example Requirements ######

pyaudio
webrtcvad
paho-mqtt